# specs/001-docusaurus-textbook/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI-Powered Physical AI & Humanoid Robotics Textbook API
  description: |
    Backend API for the interactive textbook platform providing AI-powered features:
    - RAG chatbot with citations
    - Text selection AI explanations
    - Quiz generation and attempt tracking
    - User authentication and progress tracking
    - Content personalization based on user background
    - Urdu translation with RTL support
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
servers:
  - url: https://api.textbook.example.com/api
    description: Production server
  - url: http://localhost:8000/api
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    SignupRequest:
      type: object
      required:
        - email
        - password
        - python_level
        - ai_experience
        - robotics_experience
      properties:
        email:
          type: string
          format: email
          example: "student@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "SecurePass123"
        python_level:
          type: string
          enum: [beginner, intermediate, advanced]
          example: "intermediate"
        ai_experience:
          type: string
          enum: [beginner, intermediate, advanced]
          example: "beginner"
        robotics_experience:
          type: string
          enum: [beginner, intermediate, advanced]
          example: "beginner"
        has_rtx_gpu:
          type: boolean
          default: false
        has_jetson:
          type: boolean
          default: false
        has_robot:
          type: boolean
          default: false
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (24-hour expiration)
        token_type:
          type: string
          enum: [bearer]
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
        profile:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        python_level:
          type: string
          enum: [beginner, intermediate, advanced]
        ai_experience:
          type: string
          enum: [beginner, intermediate, advanced]
        robotics_experience:
          type: string
          enum: [beginner, intermediate, advanced]
        has_rtx_gpu:
          type: boolean
        has_jetson:
          type: boolean
        has_robot:
          type: boolean

    # Chatbot Schemas
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          maxLength: 2000
          example: "What are ROS 2 topics?"
        conversation_id:
          type: string
          format: uuid
          description: Optional - for follow-up questions with context
        language:
          type: string
          enum: [en, ur]
          default: en
        chapter_context:
          type: string
          description: Current chapter ID for scoped retrieval
          example: "chapter-03-ros2-topics"
    ChatResponse:
      type: object
      properties:
        message:
          type: string
          description: AI-generated response
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        conversation_id:
          type: string
          format: uuid
        tokens_used:
          type: integer
          description: OpenAI tokens consumed
    Citation:
      type: object
      properties:
        chapter_id:
          type: string
          example: "chapter-03-ros2-topics"
        section:
          type: string
          example: "2.1 Publisher-Subscriber Pattern"
        title:
          type: string
          example: "ROS 2 Topics: Publisher-Subscriber Communication"
        relevance_score:
          type: number
          format: float
          description: Vector similarity score (0-1)
    TextSelectionRequest:
      type: object
      required:
        - selected_text
      properties:
        selected_text:
          type: string
          maxLength: 500
          example: "URDF format"
        chapter_context:
          type: string
          example: "chapter-06-gazebo-intro"
        surrounding_context:
          type: string
          maxLength: 1000
          description: Optional - surrounding paragraph for better context
    TextSelectionResponse:
      type: object
      properties:
        explanation:
          type: string
          description: AI-generated contextual explanation
        related_concepts:
          type: array
          items:
            type: string
          example: ["SDF format", "Robot kinematics", "Gazebo simulation"]
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'

    # Quiz Schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        detail:
          type: string
          example: "Email already registered"
        error_code:
          type: string
          example: "DUPLICATE_EMAIL"
paths:
...
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Register new user with background questionnaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many signup attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts (rate limited)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized (invalid or expired token)

  # Chatbot Endpoints
  /chat:
    post:
      tags:
        - Chatbot
      summary: Send message to RAG chatbot
      description: |
        Ask questions about textbook content. Responses include citations to source chapters/sections.
        Supports follow-up questions with conversation context retention.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chatbot response with citations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request (message too long, invalid conversation_id)
        '401':
          description: Unauthorized (anonymous users can chat but without conversation persistence)
        '429':
          description: Rate limited (too many requests)
        '500':
          description: OpenAI API error or vector search failure
  /chat/text-selection:
    post:
      tags:
        - Chatbot
      summary: Get AI explanation for selected text
      description: |
        Context-aware explanations for text highlighted in chapters.
        Uses chapter context and surrounding text for better understanding.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextSelectionRequest'
      responses:
        '200':
          description: Explanation generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextSelectionResponse'
        '400':
          description: Selected text too long (>500 chars)
        '429':
          description: Rate limited
  /chat/conversation/{conversation_id}:
    get:
      tags:
        - Chatbot
      summary: Retrieve conversation history
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conversation history retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                    format: uuid
                  language:
                    type: string
                    enum: [en, ur]
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        role:
                          type: string
                          enum: [user, assistant]
                        message:
                          type: string
                        citations:
                          type: array
                          items:
                            $ref: '#/components/schemas/Citation'
                        created_at:
                          type: string
                          format: date-time
        '404':
          description: Conversation not found

  # Quiz Endpoints
...
security:
  - BearerAuth: []